interview_preparation_kit/sock_merchant/main_test.go
package main_test

import (
	"bytes"
	. "dojo"
	"io/ioutil"
	"os"
	"os/exec"
	"testing"

	"github.com/stretchr/testify/assert"
)

func TestMainInputFile08ShouldBeOutputFile08(t *testing.T) {
	expectedResult, _ := ioutil.ReadFile("output/output08.txt")
	expectedResult = bytes.TrimSpace(expectedResult)
	cmd := exec.Command("go", "run", "main.go")
	cmd.Stdin, _ = os.Open("input/input08.txt")

	actualResult, _ := cmd.Output()
	actualResult = bytes.TrimSpace(actualResult)

	assert.Equal(t, expectedResult, actualResult)
}

func TestMainInputFile00ShouldBeOutputFile00(t *testing.T) {
	expectedResult, _ := ioutil.ReadFile("output/output00.txt")
	expectedResult = bytes.TrimSpace(expectedResult)
	cmd := exec.Command("go", "run", "main.go")
	cmd.Stdin, _ = os.Open("input/input00.txt")

	actualResult, _ := cmd.Output()
	actualResult = bytes.TrimSpace(actualResult)

	assert.Equal(t, expectedResult, actualResult)
}

func TestCountPairNumberInputListOfNumberShouldReturn3(t *testing.T) {
	expectedNumber := 3

	actualNumber := CountPairNumber([]int{10, 20, 20, 10, 10, 30, 50, 10, 20})

	assert.Equal(t, expectedNumber, actualNumber)
}
